Switch from MD5 to SAH512
Some Static Analysis tools consider use of MD5 a security issue, and
raise warnings when it's used. This PR remove a use of MD5 and replaces
it with SHA512, which is not considered a security issue.

The use of MD5 in this file is not (and never was) a security issue -
this is used for detecting file changes, not for any
authentication/cryptographic purposes, and so a cryptographically secure
hash is not required - MD5 is safe and fine. But SHA512 also works for
us here, so switching doesn't inconvience us much, and makes these
static analysis tools happier.